+++ src/helpers.py
@@ -1,7 +1,9 @@
 """Shared helpers used by both __main__ and api"""
 
+import os
 from typing import Any, Optional, TypedDict
 
+import httpx
 from slack_sdk.errors import SlackApiError
 
 from src.config import airtable_base, slack_client
@@ -35,9 +37,7 @@ def get_user_info(user_id: str) -> Optional[UserInfo]:
         return None
 
 
-def get_standard_channel_msg(
-    user_id: str, message_text: str
-) -> list[dict[str, Any]]:
+def get_standard_channel_msg(user_id: str, message_text: str) -> list[dict[str, Any]]:
     """Get blocks for a standard message uploaded into channel with 2 buttons"""
     return [
         {
@@ -56,10 +56,7 @@ def get_standard_channel_msg(
             "elements": [
                 {
                     "type": "mrkdwn",
-                    "text": (
-                        "Reply in this thread to send"
-                        " a response to the user"
-                    ),
+                    "text": ("Reply in this thread to send a response to the user"),
                 }
             ],
         },
@@ -82,10 +79,7 @@ def get_standard_channel_msg(
                         },
                         "text": {
                             "type": "mrkdwn",
-                            "text": (
-                                "This will mark the thread"
-                                " as complete."
-                            ),
+                            "text": ("This will mark the thread as complete."),
                         },
                         "confirm": {
                             "type": "plain_text",
@@ -146,7 +140,7 @@ def send_dm_to_user(
         )
         dm_channel: str = dm_response["channel"]["id"]
 
-        if files or reply_text == "[Shared file]":
+        if reply_text == "[Shared file]":
             return None
 
         response: dict[str, Any] = slack_client.chat_postMessage(  # type: ignore
@@ -156,9 +150,61 @@ def send_dm_to_user(
             icon_emoji=":ban:",
         )
 
+        if files:
+            download_reupload_files(files, dm_channel)
+
         return response["ts"] if response.get("ok") else None
 
     except SlackApiError as err:
         print(f"Error sending reply to user {user_id}: {err}")
         print(f"Error response: {err.response}")
         return None
+
+
+def download_reupload_files(
+    files: list[dict[str, Any]],
+    channel: str,
+    thread_ts: Optional[str] = None,
+) -> list[dict[str, Any]]:
+    """Download files, then reupload them to the target channel"""
+    reuploaded: list[dict[str, Any]] = []
+    for file in files:
+        try:
+            file_url: Optional[str] = file.get("url_private_download") or file.get(
+                "url_private"
+            )
+            if not file_url:
+                print(
+                    f"Can't really download without any url for file {file.get('name', 'unknown')}"
+                )
+                continue
+
+            headers = {"Authorization": f"Bearer {os.getenv('SLACK_BOT_TOKEN')}"}
+            response = httpx.get(file_url, headers=headers, timeout=10)
+
+            if response.status_code == 200:
+                upload_params: dict[str, Any] = {
+                    "channel": channel,
+                    "file": response.content,
+                    "filename": file.get("name", "file"),
+                    "title": file.get(
+                        "title", file.get("name", "Some file without name?")
+                    ),
+                }
+
+                if thread_ts:
+                    upload_params["thread_ts"] = thread_ts
+
+                upload_response: dict[str, Any] = slack_client.files_upload_v2(  # type: ignore
+                    **upload_params
+                )
+
+                if upload_response.get("ok"):
+                    reuploaded.append(upload_response["file"])
+                else:
+                    print(f"Failed to reupload file: {upload_response.get('error')}")
+
+        except Exception as err:  # pylint: disable=broad-except
+            print(f"Error processing file: {file.get('name', 'unknown')}: {err}")
+
+    return reuploaded
